requirements.txt
fastapi>=0.109.0
pydantic>=2.5.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
uvicorn>=0.27.0
pyyaml>=6.0.1
scikit-learn>=1.4.0
numpy>=1.26.0
httpx>=0.27.0
websockets>=12.0
# Optional: richer semantic understanding (graceful fallback if not installed)
# sentence-transformers>=2.3.0
